BX-1847 Handle Ethereum forks correctly (#762)
* BX-1847 starting refactors for tracking blockchain state

* BX-1847 expand unit tests

* BX-1847 cover all recovery paths

* BX-1847 ordered all blocks in queue

* BX-1847 fix recovery timeout

* BX-1847 remove all stale blcoks

* BX-1847 updated all the test cases

* BX-1847 fix up other unit tests

* cover some more cases

* add block stats

* BX-1847 start updating querying test cases

* BX-1847 add test coverage for doing this

* BX-1847 add logic for partial chainstate tracking

* BX-1847 update all unit tests adn get hash requesting correct

* BX-1847 fix up test cases

* BX-1847 fix up some more test cases

* BX-1847 type checking

* BX-1847 add fixme for block cleanup annotation

* BX-1847 add comments throughout tests

* BX-1847 clean up code

* BX-1847 add one more test case

+++ src/bxgateway/testing/gateway_helpers.py
@@ -41,7 +41,7 @@ def get_gateway_opts(
     block_interval=600,
     non_ssl_port: int = 9001,
     has_fully_updated_tx_service: bool = False,
-    max_block_interval: int = 10,
+    max_block_interval: float = 10,
     default_tx_quota_type: QuotaType = QuotaType.FREE_DAILY_QUOTA,
     log_level_overrides=None,
     enable_network_content_logs=False,
